refactor: extract event handlers
